Backend/requirements.txt
# ==== Core Web Framework ====
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9

# ==== Settings ====
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ==== HTTP & Async ====
httpx==0.27.0
requests==2.32.3
aiohttp==3.9.5

# ==== Image / Media ====
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4

# ==== Windows ====
python-magic-bin

# ==== Utilities ====
xmltodict==0.13.0
structlog==24.1.0
typing-extensions==4.12.1

# ==== build tool ====
setuptools>=69.5.1
wheel>=0.43.0
